test: move enum bits to their own library
