+++ json_serializable/test/integration/integration_test.dart
@@ -5,6 +5,7 @@
 import 'package:test/test.dart';
 
 import '../test_utils.dart';
+import 'json_enum_example.dart';
 import 'json_test_common.dart' show Category, Platform, StatusCode;
 import 'json_test_example.dart';
