+++ json_serializable/test/integration/json_test_example.g_any_map.dart
@@ -247,17 +247,3 @@ class PrivateConstructor {
   bool operator ==(Object other) =>
       other is PrivateConstructor && id == other.id && value == other.value;
 }
-
-@JsonEnum(alwaysCreate: true)
-enum StandAloneEnum {
-  @JsonValue('a')
-  alpha,
-  @JsonValue('b')
-  beta,
-  @JsonValue('g')
-  gamma,
-  @JsonValue('d')
-  delta,
-}
-
-Iterable<String> get standAloneEnumKeys => _$StandAloneEnumEnumMap.values;
